chore : CRA 기본 설정시 주어지는 불필요한 파일 정리
