+++ src/App.jsx
@@ -1,0 +1,8 @@
+import './App.css';
+import React from 'react';
+
+function App() {
+  return <div className="App"></div>;
+}
+
+export default App;
